abstract the buffer and indexed drawing more
