+++ Source/Engine/Systems/Renderer/Core/Meshes/MeshBufferData.h
@@ -2,7 +2,7 @@
 
 #include <memory>
 #include "Engine/SwimEngine.h"
-#include "Engine/Systems/Renderer/Vulkan/VulkanBuffer.h"
+#include "Engine/Systems/Renderer/Vulkan/Buffers/VulkanBuffer.h"
 #include "Engine/Systems/Renderer/OpenGL/OpenGLBuffer.h"
 
 namespace Engine
