+++ Source/Engine/Systems/Renderer/Vulkan/VulkanRenderer.h
@@ -7,7 +7,7 @@
 #include <unordered_map>
 #include <memory>
 #include <vector>
-#include "VulkanBuffer.h"
+#include "Buffers/VulkanBuffer.h"
 #include "Engine/Components/Transform.h"
 #include "Engine/Components/Material.h"
 #include "Engine/Systems/Renderer/Core/Meshes/Vertex.h"
@@ -18,7 +18,8 @@
 #include "VulkanSyncManager.h"
 #include "VulkanPipelineManager.h"
 #include "VulkanDescriptorManager.h"
-#include "VulkanInstanceBuffer.h"
+#include "Buffers/VulkanInstanceBuffer.h"
+#include "VulkanIndexDraw.h"
 
 // Forward declare
 struct GLFWwindow; // if we use GLFW in the future for windowing
@@ -26,16 +27,6 @@ struct GLFWwindow; // if we use GLFW in the future for windowing
 
 namespace Engine
 {
-
-	struct GpuInstanceData
-	{
-		glm::mat4 model;
-		uint32_t textureIndex; // index into bindless texture array
-		float hasTexture; // kinda scuffed flag, might be used for transparency later beyond just a "use texture" flag
-		// Pad to 16 bytes
-		float padA;
-		float padB;
-	};
 
 	class VulkanRenderer : public Machine
 	{
@@ -144,6 +135,7 @@ namespace Engine
 
 		std::unique_ptr<VulkanSyncManager> syncManager;
 		std::unique_ptr<VulkanDescriptorManager> descriptorManager;
+		std::unique_ptr<VulkanIndexDraw> indexDraw;
 
 		bool framebufferResized = false;
 
@@ -155,30 +147,13 @@ namespace Engine
 
 		VkSampler CreateSampler();
 
-		inline MeshBufferData& GetOrCreateMeshBuffers(const std::shared_ptr<Mesh>& mesh);
-
 		void UpdateUniformBuffer();
-		void UpdateInstanceBuffer();
 
 		// TODO: sampler and blend mode maps
 		VkSampler defaultSampler = VK_NULL_HANDLE; // assigned from CreateSampler during Init
 		// VkSampler activeSampler;
 
 		std::shared_ptr<Texture2D> missingTexture;
-
-		// TODO: put this somewhere else more logical (probably an instance manager class)
-		// Instance buffer manager
-		std::unique_ptr<Engine::VulkanInstanceBuffer> instanceBuffer;
-		// Stores GPU instance data to upload
-		std::vector<GpuInstanceData> cpuInstanceData;
-
-		struct MeshInstanceRange
-		{
-			uint32_t firstInstance = 0;
-			uint32_t count = 0;
-		};
-
-		std::unordered_map<std::shared_ptr<Mesh>, MeshInstanceRange> meshToInstanceOffsets;
 
 	};
