+++ Swim Engine.vcxproj
@@ -210,11 +210,13 @@ xcopy "$(ProjectDir)Source\Shaders\OpenGL" "$(TargetDir)Shaders\OpenGL" /E /I /Y
     <ClCompile Include="Source\Engine\Systems\Renderer\OpenGL\OpenGLBuffer.cpp" />
     <ClCompile Include="Source\Engine\Systems\Renderer\OpenGL\OpenGLRenderer.cpp" />
     <ClCompile Include="Source\Engine\Systems\Renderer\OpenGL\ShaderToyRendererGL.cpp" />
-    <ClCompile Include="Source\Engine\Systems\Renderer\Vulkan\VulkanBuffer.cpp" />
+    <ClCompile Include="Source\Engine\Systems\Renderer\Vulkan\Buffers\VulkanBuffer.cpp" />
+    <ClInclude Include="Source\Engine\Systems\Renderer\Vulkan\Buffers\VulkanGpuInstanceData.h" />
     <ClCompile Include="Source\Engine\Systems\Renderer\Vulkan\VulkanCommandManager.cpp" />
     <ClCompile Include="Source\Engine\Systems\Renderer\Vulkan\VulkanDescriptorManager.cpp" />
     <ClCompile Include="Source\Engine\Systems\Renderer\Vulkan\VulkanDeviceManager.cpp" />
-    <ClCompile Include="Source\Engine\Systems\Renderer\Vulkan\VulkanInstanceBuffer.cpp" />
+    <ClCompile Include="Source\Engine\Systems\Renderer\Vulkan\Buffers\VulkanInstanceBuffer.cpp" />
+    <ClCompile Include="Source\Engine\Systems\Renderer\Vulkan\VulkanIndexDraw.cpp" />
     <ClCompile Include="Source\Engine\Systems\Renderer\Vulkan\VulkanPipelineManager.cpp" />
     <ClCompile Include="Source\Engine\Systems\Renderer\Vulkan\VulkanRenderer.cpp" />
     <ClCompile Include="Source\Engine\Systems\Renderer\Vulkan\VulkanSwapChain.cpp" />
@@ -255,11 +257,12 @@ xcopy "$(ProjectDir)Source\Shaders\OpenGL" "$(TargetDir)Shaders\OpenGL" /E /I /Y
     <ClInclude Include="Source\Engine\Systems\Renderer\OpenGL\OpenGLBuffer.h" />
     <ClInclude Include="Source\Engine\Systems\Renderer\OpenGL\OpenGLRenderer.h" />
     <ClInclude Include="Source\Engine\Systems\Renderer\OpenGL\ShaderToyRendererGL.h" />
-    <ClInclude Include="Source\Engine\Systems\Renderer\Vulkan\VulkanBuffer.h" />
+    <ClInclude Include="Source\Engine\Systems\Renderer\Vulkan\Buffers\VulkanBuffer.h" />
     <ClInclude Include="Source\Engine\Systems\Renderer\Vulkan\VulkanCommandManager.h" />
     <ClInclude Include="Source\Engine\Systems\Renderer\Vulkan\VulkanDescriptorManager.h" />
     <ClInclude Include="Source\Engine\Systems\Renderer\Vulkan\VulkanDeviceManager.h" />
-    <ClInclude Include="Source\Engine\Systems\Renderer\Vulkan\VulkanInstanceBuffer.h" />
+    <ClInclude Include="Source\Engine\Systems\Renderer\Vulkan\Buffers\VulkanInstanceBuffer.h" />
+    <ClInclude Include="Source\Engine\Systems\Renderer\Vulkan\VulkanIndexDraw.h" />
     <ClInclude Include="Source\Engine\Systems\Renderer\Vulkan\VulkanPipelineManager.h" />
     <ClInclude Include="Source\Engine\Systems\Renderer\Vulkan\VulkanRenderer.h" />
     <ClInclude Include="Source\Engine\Systems\Renderer\Vulkan\VulkanSwapChain.h" />
